feat(DepartmentNoticeRepository): 사용하지 않는 DepartmentNoticeRepository 제거
