--- src/main/java/com/kustacks/kuring/notice/domain/DepartmentNoticeRepository.java
@@ -1,6 +1,0 @@
-package com.kustacks.kuring.notice.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DepartmentNoticeRepository extends JpaRepository<DepartmentNotice, Long>, DepartmentNoticeQueryRepository {
-}
