+++ src/main/java/com/kustacks/kuring/notice/domain/NoticeQueryRepositoryImpl.java
@@ -14,6 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
+import static com.kustacks.kuring.notice.domain.QDepartmentNotice.departmentNotice;
 import static com.kustacks.kuring.notice.domain.QNotice.notice;
 
 @RequiredArgsConstructor
@@ -72,6 +73,82 @@ public class NoticeQueryRepositoryImpl implements NoticeQueryRepository {
                 .delete(notice)
                 .where(notice.categoryName.eq(categoryName)
                         .and(notice.articleId.in(articleIds)))
+                .execute();
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public List<Integer> findImportantArticleIdsByDepartment(DepartmentName departmentName) {
+        return queryFactory
+                .select(departmentNotice.articleId.castToNum(Integer.class))
+                .from(departmentNotice)
+                .where(departmentNotice.departmentName.eq(departmentName)
+                        .and(departmentNotice.important.eq(true)))
+                .orderBy(departmentNotice.articleId.castToNum(Integer.class).asc())
+                .fetch();
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public List<Integer> findNormalArticleIdsByDepartment(DepartmentName departmentName) {
+        return queryFactory
+                .select(departmentNotice.articleId.castToNum(Integer.class))
+                .from(departmentNotice)
+                .where(departmentNotice.departmentName.eq(departmentName)
+                        .and(departmentNotice.important.eq(false)))
+                .orderBy(departmentNotice.articleId.castToNum(Integer.class).asc())
+                .fetch();
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public List<NoticeDto> findImportantNoticesByDepartment(DepartmentName departmentName) {
+        return queryFactory
+                .select(new QNoticeDto(
+                        departmentNotice.articleId,
+                        departmentNotice.postedDate,
+                        departmentNotice.url.value,
+                        departmentNotice.subject,
+                        departmentNotice.categoryName.stringValue().toLowerCase(),
+                        departmentNotice.important))
+                .from(departmentNotice)
+                .where(departmentNotice.departmentName.eq(departmentName)
+                        .and(departmentNotice.important.isTrue()))
+                .orderBy(departmentNotice.postedDate.desc())
+                .fetch();
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public List<NoticeDto> findNormalNoticesByDepartmentWithOffset(DepartmentName departmentName, Pageable pageable) {
+        return queryFactory
+                .select(new QNoticeDto(
+                        departmentNotice.articleId,
+                        departmentNotice.postedDate,
+                        departmentNotice.url.value,
+                        departmentNotice.subject,
+                        departmentNotice.categoryName.stringValue().toLowerCase(),
+                        departmentNotice.important))
+                .from(departmentNotice)
+                .where(departmentNotice.departmentName.eq(departmentName)
+                        .and(departmentNotice.important.isFalse()))
+                .offset(pageable.getOffset())
+                .limit(pageable.getPageSize())
+                .orderBy(departmentNotice.postedDate.desc())
+                .fetch();
+    }
+
+    @Transactional
+    @Override
+    public void deleteAllByIdsAndDepartment(DepartmentName departmentName, List<String> articleIds) {
+        if(articleIds.isEmpty()) {
+            return;
+        }
+
+        queryFactory
+                .delete(departmentNotice)
+                .where(departmentNotice.departmentName.eq(departmentName)
+                        .and(departmentNotice.articleId.in(articleIds)))
                 .execute();
     }
