+++ src/test/java/com/kustacks/kuring/tool/DatabaseConfigurator.java
@@ -22,7 +22,6 @@ import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Component
@@ -39,7 +38,6 @@ public class DatabaseConfigurator implements InitializingBean {
     private final UserRepository userRepository;
     private final StaffRepository staffRepository;
     private final AdminRepository adminRepository;
-    private final DepartmentNoticeRepository departmentNoticeRepository;
     private final DataSource dataSource;
     private final JdbcTemplate jdbcTemplate;
     private final PasswordEncoder passwordEncoder;
@@ -48,13 +46,11 @@ public class DatabaseConfigurator implements InitializingBean {
 
     public DatabaseConfigurator(NoticeRepository noticeRepository, UserRepository userRepository,
                                 StaffRepository staffRepository, AdminRepository adminRepository,
-                                DepartmentNoticeRepository departmentNoticeRepository,
                                 DataSource dataSource, JdbcTemplate jdbcTemplate, PasswordEncoder passwordEncoder) {
         this.noticeRepository = noticeRepository;
         this.userRepository = userRepository;
         this.staffRepository = staffRepository;
         this.adminRepository = adminRepository;
-        this.departmentNoticeRepository = departmentNoticeRepository;
         this.dataSource = dataSource;
         this.jdbcTemplate = jdbcTemplate;
         this.passwordEncoder = passwordEncoder;
@@ -154,10 +150,10 @@ public class DatabaseConfigurator implements InitializingBean {
         noticeRepository.saveAll(noticeList);
 
         List<DepartmentNotice> importantDeptNotices = buildDepartmentNotice(7, DepartmentName.COMPUTER, CategoryName.DEPARTMENT, true);
-        departmentNoticeRepository.saveAll(importantDeptNotices);
+        noticeRepository.saveAll(importantDeptNotices);
 
         List<DepartmentNotice> normalDeptNotices = buildDepartmentNotice(5, DepartmentName.COMPUTER, CategoryName.DEPARTMENT, false);
-        departmentNoticeRepository.saveAll(normalDeptNotices);
+        noticeRepository.saveAll(normalDeptNotices);
     }
 
     private void initStaff() {
