+++ src/test/java/com/kustacks/kuring/worker/update/DepartmentNoticeUpdaterTest.java
@@ -2,7 +2,8 @@ package com.kustacks.kuring.worker.update;
 
 import com.kustacks.kuring.message.firebase.FirebaseService;
 import com.kustacks.kuring.notice.domain.DepartmentNotice;
-import com.kustacks.kuring.notice.domain.DepartmentNoticeRepository;
+import com.kustacks.kuring.notice.domain.Notice;
+import com.kustacks.kuring.notice.domain.NoticeRepository;
 import com.kustacks.kuring.worker.scrap.DepartmentNoticeScraperTemplate;
 import com.kustacks.kuring.worker.scrap.dto.ComplexNoticeFormatDto;
 import com.kustacks.kuring.worker.update.notice.DepartmentNoticeUpdater;
@@ -20,6 +21,7 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.Mockito.doNothing;
@@ -46,7 +48,7 @@ class DepartmentNoticeUpdaterTest {
     ThreadPoolTaskExecutor noticeUpdaterThreadTaskExecutor;
 
     @Autowired
-    DepartmentNoticeRepository departmentNoticeRepository;
+    NoticeRepository noticeRepository;
 
     @DisplayName("학과별 공지 업데이트 테스트")
     @Test
@@ -60,8 +62,11 @@ class DepartmentNoticeUpdaterTest {
         noticeUpdaterThreadTaskExecutor.getThreadPoolExecutor().awaitTermination(2, TimeUnit.SECONDS);
 
         // then
-        List<DepartmentNotice> notices = departmentNoticeRepository.findAll();
-        assertThat(notices).hasSize(3720);
+        List<Notice> notices = noticeRepository.findAll();
+        assertAll(
+                () -> assertThat(notices).hasSize(3720),
+                () -> assertThat(notices.get(0)).isExactlyInstanceOf(DepartmentNotice.class)
+        );
     }
 
     private static List<ComplexNoticeFormatDto> createDepartmentNoticesFixture() {
